requirements.txt

streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
google-generativeai==0.3.2
scikit-learn==1.4.0
joblib==1.3.2
